Crate packet_sniffer

Modules

address_port

Module defining the AddressPort struct, which represents a network address:port pair.

args

Module defining the Args struct used for command line options.

report_info

Module defining the ReportInfo struct, useful to format the output report file and to keep track of statistics about the sniffed traffic.

thread_parse_packets_functions

Module containing functions executed by the thread in charge of parsing sniffed packets and inserting them in the shared map.

thread_write_report_functions

Module containing functions executed by the thread in charge of updating the output report every interval seconds, with interval specified by the user through the -i command line option.

Enums

Status

Functions

is_valid_network_layer

Checks if the provided network_layer equals “ipv6” or “ipv4” or “no filter”.

is_valid_transport_layer

Checks if the provided transport_layer equals “tcp” or “udp” or “no filter”.

main

Entry point of application execution.

print_device_list

Prints the list of available network adapters’ names and addresses.

retrieve_device

Given the name of the desired network adapter, this function returns an Option<Device> which contains the corresponding Device struct if the provided network adapter exists or a None value if it doesn’t exist.

set_status_by_key