Struct packet_sniffer::command_line_options::Args

pub struct Args {
    pub adapter: String,
    pub device_list: bool,
    pub highest_port: u16,
    pub interval: u64,
    pub lowest_port: u16,
    pub minimum_packets: u32,
    pub network_layer_filter: String,
    pub output_file: String,
    pub transport_layer_filter: String,
}

Struct for command line arguments, to be parsed using clap.

Fields
adapter: String

Name of the network adapter to be inspected, if omitted the default adapter is chosen.

device_list: bool

Prints the list of available devices. Immediately terminates the program.

highest_port: u16

Sets the maximum port value to be considered, if omitted there is not ports higher bound.

interval: u64

Sets the interval of time between report updates (value in seconds).

lowest_port: u16

Sets the minimum port value to be considered, if omitted there is not ports lower bound.

minimum_packets: u32

Sets the minimum value of transited packets for an address:port to be printed in the report.

network_layer_filter: String

Filters packets on the basis of the IP version address (IPv4 or IPv6).

output_file: String

Name of output file to contain the textual report, if omitted a default file is chosen.

transport_layer_filter: String

Filters packets on the basis of the transport layer protocol (TCP or UDP).