Function packet_sniffer::thread_parse_packets_functions::analyze_network_header

fn analyze_network_header(
    network_header: Option<IpHeader>, 
    exchanged_bytes: &mut u32, 
    network_layer: &mut String, 
    address1: &mut String, 
    address2: &mut String, 
    skip_packet: &mut bool
)

This function analyzes the network layer header passed as parameter and updates variables passed by reference on the basis of the packet header content.

Arguments

network_header - An Option containing the IpHeader enum obtained through etherparse.

exchanged_bytes - Parameter initialized with a default value; it is passed by reference and it will be overwritten with the IP payload dimension.

network_layer - Parameter initialized with a default value; it is passed by reference and it will be overwritten with a String representing the IP version (“ipv4” or “ipv6”).

address1 - Parameter initialized with a default value; it is passed by reference and will be overwritten with the source IP address of the packet.

address2 - Parameter initialized with a default value; it is passed by reference and will be overwritten with the destination IP address of the packet.

skip_packet - Boolean flag initialized with a false value; it is passed by reference and will be overwritten to true if the network_header is invalid; in this case the packet will not be considered.