Function packet_sniffer::thread_parse_packets_functions::analyze_transport_header

fn analyze_transport_header(
    transport_header: Option<TransportHeader>, 
    transport_layer: &mut String, 
    port1: &mut u16, 
    port2: &mut u16, 
    application_protocol_1: &mut Option<AppProtocol>, 
    application_protocol_2: &mut Option<AppProtocol>, 
    application_protocols: &mut HashSet<AppProtocol>, 
    transport_protocol: &mut TransProtocol, 
    skip_packet: &mut bool
)

This function analyzes the transport layer header passed as parameter and updates variables passed by reference on the basis of the packet header content.

Arguments

transport_header - An Option containing the TransportHeader enum obtained through etherparse.

transport_layer - Parameter initialized with a default value; it is passed by reference and it will be overwritten with a String representing the transport layer protocol (“tcp” or “udp”).

port1 - Parameter initialized with a default value; it is passed by reference and will be overwritten with the source port of the packet.

port2 - Parameter initialized with a default value; it is passed by reference and will be overwritten with the destination port of the packet.

application_protocol_1 - Parameter initialized with a None value; it is passed by reference and will be overwritten with the application layer protocol obtained from the source port.

application_protocol_2 - Parameter initialized with a None value; it is passed by reference and will be overwritten with the application layer protocol obtained from the destination port.

application_protocols - Parameter initialized with a default value; it is passed by reference and will be overwritten with a set containing the application layer protocols obtained from the source and destination ports.

transport_protocol - Parameter initialized with a default value; it is passed by reference and will be overwritten with the observed transport layer protocol.

skip_packet - Boolean flag initialized with a false value; it is passed by reference and will be overwritten to true if the transport_header is invalid; in this case the packet will not be considered.