Function packet_sniffer::thread_parse_packets_functions::from_port_to_application_protocol

fn from_port_to_application_protocol(port: u16) -> Option<AppProtocol>

Given an integer in the range 0..=65535, this function returns an Option<AppProtocol> containing the respective application protocol represented by a value of the AppProtocol enum. Only the most common application layer protocols are considered; if a unknown port number is provided, this function returns None.

Arguments

port - An integer representing the transport layer port to be mapped to an application layer protocol.

Examples

let x = from_port_to_application_protocol(25);
//Simple Mail Transfer Protocol
assert_eq!(x, Option::Some(AppProtocol::SMTP));

let y = from_port_to_application_protocol(1999);
//Unknown port-to-protocol mapping
assert_eq!(y, Option::None);