Function packet_sniffer::thread_parse_packets_functions::modify_or_insert_destination_in_map

fn modify_or_insert_destination_in_map(
    mutex_map: Arc<Mutex<HashMap<AddressPort, ReportInfo>>>, 
    key: AddressPort, 
    exchanged_bytes: u32, 
    transport_protocol: TransProtocol, 
    application_protocol_1: Option<AppProtocol>, 
    application_protocol_2: Option<AppProtocol>, 
    application_protocols: HashSet<AppProtocol>
)

Function to insert the destination of a packet into the shared map containing the analyzed traffic.

Arguments

mutex_map - Mutex to permit exclusive access to the shared map containing the parsed packets.

key - An AddressPort element representing the destination of the packet. It corresponds to the map key part.

exchanged_bytes - IP payload dimension of the observed packet.

transport_protocol - Transport layer protocol carried by the observed packet.

application_protocol_1 - Application layer protocol obtained from the source port.

application_protocol_2 - Application layer protocol obtained from the destination port.

application_protocols - Set containing the application layer protocols obtained from the source and destination ports.