Function packet_sniffer::thread_parse_packets_functions::parse_packets_loop

pub fn parse_packets_loop(
    cap: Capture<Active>, 
    lowest_port: u16, 
    highest_port: u16, 
    network_layer_filter: String, 
    transport_layer_filter: String, 
    mutex_map: Arc<Mutex<HashMap<AddressPort, ReportInfo>>>, 
    status_pair: Arc<(Mutex<Status>, Condvar)>
)

The calling thread enters in a loop in which it waits for network packets, parses them according to the user specified filters, and inserts them into the shared map variable.

Arguments

cap - Initialized Capture handle to interface with the network traffic

lowest_port - The lowest port number to be considered in the report. Specified by the user through the -l option.

highest_port - The highest port number to be considered in the report. Specified by the user through the -h option.

network_layer - A String representing the IP version to be filtered. Specified by the user through the -n option.

transport_layer - A String representing the transport protocol to be filtered. Specified by the user through the -t option.

mutex_map - Mutex to permit exclusive access to the shared variable in which the parsed packets are inserted.