Function packet_sniffer::write_report_functions::write_report_file_header

fn write_report_file_header(
    output: File, 
    device_name: String, 
    first_timestamp: String, 
    times_report_updated: i32, 
    interval: u64, 
    lowest_port: u16, 
    highest_port: u16, 
    min_packets: u32, 
    network_layer: String, 
    transport_layer: String
)

Writes the output report file header, which contains useful info about the sniffing process.

Arguments

output_file - A String representing the output report file name. Specified by the user through the -o option.

device_name - A String representing the name of th network adapter to be sniffed. Specified by the user through the -a option.

first_timestamp - A not formatted String representing the initial timestamp of the sniffing process.

times_report_updated - An integer representing the amount of times the report has been updated.

interval - Frequency of report updates (value in seconds). Specified by the user through the -i option.

lowest_port - The lowest port number to be considered in the report. Specified by the user through the -l option.

highest_port - The highest port number to be considered in the report. Specified by the user through the -h option.

min_packets - Minimum number of packets for an address:port pair to be considered in the report. Specified by the user through the -m option.

network_layer - A String representing the IP version to be filtered. Specified by the user through the -n option.

transport_layer - A String representing the transport protocol to be filtered. Specified by the user through the -t option.

Examples

An example of output report file header is reported below.

<><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><>
<><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><>
<><>
<><>		Packets are sniffed from adapter 'en0'
<><>
<><>		Report updates info
<><>			Report start time: 15/08/2022 15:39:07
<><>			Report last update: 15/08/2022 15:39:52
<><>			Report update frequency: every 5 seconds
<><>			Number of times report was updated: 9
<><>
<><>		Filters
<><>			Considering only address:port pairs featured by more than 500 packets
<><>			Considering both IPv4 and IPv6 packets
<><>			Considering only packets exchanged with TCP
<><>			Considering only port number 443
<><>
<><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><>
<><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><><>