Struct packet_sniffer::command_line_options::Args

pub struct Args {
    pub adapter: String,
    pub device_list: bool,
    pub highest_port: u16,
    pub interval: u64,
    pub lowest_port: u16,
    pub minimum_packets: u32,
    pub network_layer_filter: String,
    pub output_file: String,
    pub transport_layer_filter: String,
}

Struct for command line arguments, to be parsed using clap.

Fields
adapter: String

-a, --adapter

Name of the network adapter to be inspected, if omitted the default adapter is chosen.

If a non-existing adapter is provided, the application raises an error and terminates.

This option must be followed by a textual value.

device_list: bool

-d, --device-list

Prints list of the available network interfaces.

Immediately terminates the program.

This option does not need to be followed by a value.

highest_port: u16

-h, --highest-port

Sets the maximum port value to be considered, if omitted there is not ports higher bound.

If the highest-port provided value is lower than the lowest-port provided value, the application raises an error and terminates.

This option must be followed by an integer value between 0 and 65535.

interval: u64

-i, --interval

Sets the interval of time between report updates (value in seconds).

This option must be followed by a positive integer value.

lowest_port: u16

-l, --lowest-port

Sets the lowest port value to be considered, if omitted there is not ports lower bound.

If the lowest-port provided value is lower than the highest-port provided value, the application raises an error and terminates.

This option must be followed by an integer value between 0 and 65535.

minimum_packets: u32

-m, --minimum-packets

Sets the minimum value of transited packets for an address:port to be printed in the report.

This option must be followed by a positive integer value.

network_layer_filter: String

-n, --network-layer-filter

Filters packets on the basis of the IP version address (IPv4 or IPv6).

This option must be followed by a textual value.

output_file: String

-o, --output-file

Name of output file to contain the textual report, if omitted a default file is chosen.

This option must be followed by a textual value.

transport_layer_filter: String

-t, --transport-layer-filter

Filters packets on the basis of the transport layer protocol (TCP or UDP).

This option must be followed by a textual value.